feat: hardware capability layer, precision enum, working fp16 path
Capability probing was scattered across four call sites that each ran
torch.cuda.is_bf16_supported() inline. That call counts *emulated* bf16, so
it cannot answer "is bf16 native here?" — and the fp16 escape hatch it
implied was not actually trainable.

Hardware layer
- ops/hardware/cuda.py: runtime probes (compute capability, native bf16,
  flash-attn 2, memory, describe()). Criterion is sm >= 80, not
  is_bf16_supported().
- ops/hardware/device.py: dispatch entries for the new probes; backend
  import made lazy, per that subtree's contract.
- ops/hardware/precision.py: new, backend-independent policy —
  resolve_precision / to_torch_dtype / to_storage_dtype / to_vllm_dtype_name
  / needs_autocast / needs_grad_scaler / needs_fp32_master_weights.
- resolve_attn_implementation now gates flash-attn 2 on sm_80+; it imports
  fine on Turing but crashes in forward.
- main.py logs the resolved capabilities and precision before Ray starts.

Config surface
- precision: auto|bf16|fp16|fp32 replaces the use_bf16 boolean, which could
  express neither fp32 nor "let the framework decide". use_bf16 still maps
  through flat_config.py with a DeprecationWarning; precision wins when both
  are present.
- auto resolves to bf16 on every CUDA GPU and fp32 without one. On sm_75,
  emulated bf16 measured 2026 ms / 20.16 GiB against 2831 ms / 22.96 GiB for
  the only correct fp16 path, so fp16 is an explicit opt-in, not a fallback.

fp16 correctness
- fp32 master weights under fp16: FSDP2 asserts a uniform original dtype per
  sharded unit (a LoRA layer holds frozen base and trainable adapter in one),
  so storage dtype is uniform fp32 for a trainable group while
  MixedPrecisionPolicy casts the all-gathered copy to fp16 for compute.
  Frozen groups stay at compute dtype — no memory regression.
- GradScaler on ParallelModelGroup: unscale_ before clip_grad_norm_, state
  saved to scaler.pt and restored on resume, grad_overflow surfaced in
  step_info so a skipped step is visible.
- Engine-level torch.autocast under fp16. Blanket param casting alone
  produces NaN gradients in Qwen3.5's gated linear attention from layer 14
  down, at any loss scale, while the loss stays finite.
- DistCrossEntropyLoss upcasts to fp32, matching the chunked path.

tests/test_precision.py covers the resolution table, the use_bf16 migration,
the storage/compute dtype split and the fp16 clip-to-zero regression.

+++ razordl/ops/loss/distributed.py
@@ -16,7 +16,12 @@ class DistCrossEntropyLoss(torch.nn.Module):
         self.ce = torch.nn.CrossEntropyLoss(ignore_index=ignore_index, reduction="none")
 
     def forward(self, logits, labels):
-        ce_loss = self.ce(logits, labels)
+        # Upcast to fp32 before the softmax and the sum, matching what the
+        # chunked path in presets/sft/workgroup.py already does. In fp16 a
+        # per-token loss sum over a few thousand tokens overflows 65504 to inf;
+        # even below that the reduction loses precision. Costs one fp32 copy of
+        # the logits -- with a large vocabulary, prefer chunked_loss.
+        ce_loss = self.ce(logits.float(), labels)
         valid_tokens = (labels != self.ignore_index).sum().item()
         batch_valid_tokens = distributed_token_count(valid_tokens)
         return ce_loss.sum() / max(batch_valid_tokens, 1)
